--HG--
branch : 2 - simple

+++ sources/count paths.cpp
@@ -14,7 +14,7 @@ uint count_paths(bool *map , uint columns , uint pi , uint pj , uint ti , uint t
 	if(*(map+pi*columns+pj))
 		return 0;
 
-	unsigned int D = abs(pi-ti) + abs(pj-tj);
+	unsigned int D = abs((int)(pi-ti)) + abs((int)(pj-tj));
 	if(D > steps || (steps-D) & 0x1)
 		return 0;
 	
@@ -35,7 +35,7 @@ uint count_paths(bool *map , uint columns , uint pi , uint pj , uint ti , uint t
 
 static uint search4way(uint pi , uint pj , uint steps)															// search4way
 {
-	if(*(map+pi*columns+pj) || steps < abs(pi-ti) + abs(pj-tj))
+	if(*(map+pi*columns+pj) || steps < (uint)(abs((int)(pi-ti)) + abs((int)(pj-tj))))
 		return 0;
 	if(steps--)
 		return search4way(pi+1,pj,steps) + search4way(pi,pj+1,steps) +
